Move workflow upload/download to their own components
