+++ src/WorkflowPanel.tsx
@@ -1,27 +1,18 @@
-import { useRef, useState } from 'react'
+import { useState } from 'react'
 import { TextPanel } from './TextPanel'
-import { useWorkflow } from './store'
 import { VisualPanel } from './VisualPanel'
+import { WorkflowDownload } from './WorkflowDownload'
+import { WorkflowUpload } from './WorkflowUpload'
 
 type ITab = 'text' | 'visual'
 
 export const WorkflowPanel = (): JSX.Element => {
-  const uploadRef = useRef<HTMLInputElement>(null)
   const [tab, setTab] = useState<ITab>('visual')
-  const { loadWorkflowArchive, save } = useWorkflow()
+
   const selectedPanel = tab === 'visual' ? <VisualPanel /> : <TextPanel />
   const visualTabStyle = { fontWeight: tab === 'visual' ? 'bold' : 'normal' }
   const textTabStyle = { fontWeight: tab === 'text' ? 'bold' : 'normal' }
 
-  async function uploadWorkflow (event: React.ChangeEvent<HTMLInputElement>): Promise<void> {
-    if (event.target.files == null) {
-      return
-    }
-    const file = event.target.files[0]
-    const url = URL.createObjectURL(file)
-    await loadWorkflowArchive(url)
-    URL.revokeObjectURL(url)
-  }
   return (
     <fieldset>
       <legend>Workflow</legend>
@@ -31,19 +22,10 @@ export const WorkflowPanel = (): JSX.Element => {
             <button className='btn btn-light' style={visualTabStyle} onClick={() => setTab('visual')}>Visual</button>
             <button className='btn btn-light' style={textTabStyle} onClick={() => setTab('text')}>Text</button>
           </div>
-          <button className='btn btn-link' onClick={() => uploadRef.current?.click()}>
-            Upload
-            <input
-              type='file'
-              accept='application/zip,.zip'
-              onChange={uploadWorkflow}
-              ref={uploadRef}
-              style={{ opacity: 0, width: 0, height: 0 }}
-            />
-          </button>
+          <WorkflowUpload />
           {selectedPanel}
         </div>
-        <button className='btn btn-primary' onClick={save}>Download archive</button>
+        <WorkflowDownload />
       </div>
     </fieldset>
   )
